Really looks like water! Fixed up constants and some force calculations, now trying time integration.

+++ 594G Final Project/FIELD_2D.cpp
@@ -11,7 +11,7 @@ FIELD_2D::FIELD_2D() :
 FIELD_2D::FIELD_2D(int xRes, int yRes) : 
   _xRes(xRes), _yRes(yRes)
 {
-  _data = new float[_xRes * _yRes];
+  _data = new double[_xRes * _yRes];
   clear();
 }
 
@@ -34,7 +34,7 @@ void FIELD_2D::resize(int xRes, int yRes) {
   if (_data) delete[] _data;
   _xRes = xRes;
   _yRes = yRes;
-  _data = new float[_xRes * _yRes];
+  _data = new double[_xRes * _yRes];
   clear();
 }
 
@@ -49,7 +49,7 @@ void FIELD_2D::clear() {
 ///////////////////////////////////////////////////////////////////////
 // scale the field by a constant
 ///////////////////////////////////////////////////////////////////////
-FIELD_2D& FIELD_2D::operator*=(const float& scalar)
+FIELD_2D& FIELD_2D::operator*=(const double& scalar)
 {
   for (int x = 0; x < _xRes * _yRes; x++)
     _data[x] *= scalar;
@@ -74,9 +74,9 @@ FIELD_2D& FIELD_2D::operator=(const FIELD_2D& field)
 ///////////////////////////////////////////////////////////////////////
 // squared sum of entries
 ///////////////////////////////////////////////////////////////////////
-float FIELD_2D::squaredSum() const
+double FIELD_2D::squaredSum() const
 {
-  float final = 0.0;
+  double final = 0.0;
   for (int x = 0; x < _xRes * _yRes; x++)
     final += _data[x] * _data[x];
 
@@ -86,12 +86,12 @@ float FIELD_2D::squaredSum() const
 ///////////////////////////////////////////////////////////////////////
 // dot product with another field
 ///////////////////////////////////////////////////////////////////////
-float FIELD_2D::dotProduct(FIELD_2D& field) const
+double FIELD_2D::dotProduct(FIELD_2D& field) const
 {
   assert(field.xRes() == _xRes);
   assert(field.yRes() == _yRes);
 
-  float final = 0.0;
+  double final = 0.0;
   for (int x = 0; x < _xRes * _yRes; x++)
     final += _data[x] * field._data[x];
 
@@ -130,7 +130,7 @@ FIELD_2D& FIELD_2D::operator+=(const FIELD_2D& field)
 // axpy operation with another field
 // this = this + scalar * field
 ///////////////////////////////////////////////////////////////////////
-void FIELD_2D::axpy(const float& scalar, const FIELD_2D& field)
+void FIELD_2D::axpy(const double& scalar, const FIELD_2D& field)
 {
   assert(field.xRes() == _xRes);
   assert(field.yRes() == _yRes);
